refactor: Split `utils`  for `members` and `rooms` util files
