+++ src/components/SidebarActions/DirectMessageModal.tsx
@@ -5,7 +5,6 @@ import {IoCheckmark, IoCloseCircle, IoCopyOutline} from "react-icons/io5"
 import Input from "../Input"
 import UserProfile from "../UserProfile"
 import {
-  getDirectRoomsIds,
   getPartnerUserIdFromRoomDirect,
   normalizeName,
   stringToColor,
@@ -15,6 +14,7 @@ import useConnection from "@/hooks/matrix/useConnection"
 import useInvitationLink from "@/hooks/matrix/useInvitationLink"
 import useUsersSearch from "@/hooks/matrix/useUserSearch"
 import {useSidebarModalActiveStore} from "./useSidebarActions"
+import {getDirectRoomsIds} from "@/utils/rooms"
 
 type DirectChatRecentProps = {
   userId: string
@@ -50,7 +50,7 @@ const DirectMessageModal: FC = () => {
   const {clearActiveSidebarModal} = useSidebarModalActiveStore()
   const [userId, setUserId] = useState<string | null>(null)
   const [directChats, setDirectChats] = useState<DirectChatRecentProps[]>([])
-  const {query, setQuery, results} = useUsersSearch(client)
+  const {setQuery, results} = useUsersSearch(client)
 
   const {invitationLink, isLinkCopied, copyToClipboard} =
     useInvitationLink(userId)
@@ -106,7 +106,6 @@ const DirectMessageModal: FC = () => {
 
       <Input
         className="w-full"
-        initialValue={query}
         onValueChange={setQuery}
         placeholder="Enter name or username"
       />
